+++ python/coral_usb/__init__.py
@@ -1,4 +1,6 @@
+from .posenet.pose_engine import PoseEngine  # NOQA
+
 from .detector_base import EdgeTPUDetectorBase  # NOQA
 from .face_detector import EdgeTPUFaceDetector  # NOQA
+from .human_pose_estimator import EdgeTPUHumanPoseEstimator  # NOQA
 from .object_detector import EdgeTPUObjectDetector  # NOQA
-from .posenet.pose_engine import PoseEngine  # NOQA
